Add staff of light special attack
